Chore(evm-bridge): refactor
