+++ evm-utils/evm-bridge/src/main.rs
@@ -1,11 +1,12 @@
 mod middleware;
 mod pool;
+mod rpc_client;
 
 use log::*;
 use std::future::ready;
 use std::str::FromStr;
-use std::sync::{Arc, atomic::{AtomicU64, Ordering}};
-use std::time::{Duration, Instant};
+use std::sync::Arc;
+use std::time::Duration;
 use std::{
     collections::{HashMap, HashSet},
     net::SocketAddr,
@@ -13,8 +14,8 @@ use std::{
 
 use evm_rpc::bridge::BridgeERPC;
 use evm_rpc::chain::ChainERPC;
-use evm_rpc::general::GeneralERPC;
 use evm_rpc::error::{Error, *};
+use evm_rpc::general::GeneralERPC;
 use evm_rpc::*;
 use evm_state::*;
 use sha3::{Digest, Keccak256};
@@ -23,31 +24,20 @@ use jsonrpc_core::BoxFuture;
 use jsonrpc_http_server::jsonrpc_core::*;
 use jsonrpc_http_server::*;
 
-use reqwest::{
-    header::{CONTENT_TYPE, RETRY_AFTER},
-    StatusCode,
-};
-
-use serde::Deserialize;
 use serde_json::json;
 use snafu::ResultExt;
 
 use derivative::*;
 use solana_evm_loader_program::scope::*;
 use solana_sdk::{
-    clock::{DEFAULT_MS_PER_SLOT, MS_PER_TICK, Slot}, commitment_config::CommitmentConfig,
-    fee_calculator::{DEFAULT_TARGET_LAMPORTS_PER_SIGNATURE, FeeCalculator}, hash::Hash,
-    pubkey::Pubkey, signature::Signature, signers::Signers,
-    transaction::{TransactionError, uses_durable_nonce},
+    clock::MS_PER_TICK, fee_calculator::DEFAULT_TARGET_LAMPORTS_PER_SIGNATURE, pubkey::Pubkey,
+    signers::Signers, transaction::TransactionError,
 };
-use solana_transaction_status::{TransactionStatus, UiTransactionEncoding};
 
 use solana_client::{
-    client_error::{ClientError, ClientErrorKind, Result as ClientResult},
-    rpc_client::serialize_encode_transaction,
+    client_error::{ClientError, ClientErrorKind},
     rpc_config::*,
-    rpc_custom_error,
-    rpc_request::{RpcError, RpcRequest, RpcResponseErrorData},
+    rpc_request::{RpcRequest, RpcResponseErrorData},
     rpc_response::Response as RpcResponse,
     rpc_response::*,
 };
@@ -68,6 +58,7 @@ use pool::{
     worker_cleaner, worker_deploy, worker_signature_checker, EthPool, PooledTransaction,
     SystemClock,
 };
+use rpc_client::AsyncRpcClient;
 
 use rlp::Encodable;
 use secp256k1::Message;
@@ -195,478 +186,6 @@ macro_rules! proxy_evm_rpc {
         }
     )
 
-}
-
-#[derive(Deserialize, Debug)]
-struct RpcErrorObject {
-    code: i64,
-    message: String,
-    #[serde(default)]
-    data: Value,
-}
-
-pub struct AsyncRpcClient {
-    client: Arc<reqwest::Client>,
-    url: String,
-    request_id: AtomicU64,
-}
-
-impl AsyncRpcClient {
-    pub fn new(url: String) -> Self {
-        Self::new_with_timeout(url, Duration::from_secs(30))
-    }
-
-    pub fn new_with_timeout(url: String, timeout: Duration) -> Self {
-        let client = Arc::new(
-            reqwest::Client::builder()
-                .timeout(timeout)
-                .build()
-                .expect("build rpc client")
-        );
-
-        Self {
-            client,
-            url,
-            request_id: AtomicU64::new(0),
-        }
-    }
-
-    pub async fn send_request(&self, request: RpcRequest, params: Value) -> ClientResult<Value> {
-        let request_id = self.request_id.fetch_add(1, Ordering::Relaxed);
-        let request_json = request.build_request_json(request_id, params).to_string();
-        match self._send_request(request_json).await {
-            Ok(response) => {
-                let json: Value = response.json().await?;
-                if json["error"].is_object() {
-                    match serde_json::from_value::<RpcErrorObject>(json["error"].clone())
-                    {
-                        Ok(rpc_error_object) => {
-                            let data = match rpc_error_object.code {
-                                rpc_custom_error::JSON_RPC_SERVER_ERROR_SEND_TRANSACTION_PREFLIGHT_FAILURE => {
-                                    match serde_json::from_value::<RpcSimulateTransactionResult>(json["error"]["data"].clone()) {
-                                        Ok(data) => RpcResponseErrorData::SendTransactionPreflightFailure(data),
-                                        Err(err) => {
-                                            debug!("Failed to deserialize RpcSimulateTransactionResult: {:?}", err);
-                                            RpcResponseErrorData::Empty
-                                        }
-                                    }
-                                },
-                                rpc_custom_error::JSON_RPC_SERVER_ERROR_NODE_UNHEALTHY => {
-                                    match serde_json::from_value::<rpc_custom_error::NodeUnhealthyErrorData>(json["error"]["data"].clone()) {
-                                        Ok(rpc_custom_error::NodeUnhealthyErrorData {num_slots_behind}) => RpcResponseErrorData::NodeUnhealthy {num_slots_behind},
-                                        Err(_err) => {
-                                            RpcResponseErrorData::Empty
-                                        }
-                                    }
-                                },
-                                _ => RpcResponseErrorData::Empty
-                            };
-
-                            Err(RpcError::RpcResponseError {
-                                code: rpc_error_object.code,
-                                message: rpc_error_object.message,
-                                data,
-                                original_err: json["error"]["data"].clone(),
-                            }
-                                .into())
-                        }
-                        Err(err) => Err(RpcError::RpcRequestError(format!(
-                            "Failed to deserialize RPC error response: {} [{}]",
-                            serde_json::to_string(&json["error"]).unwrap(),
-                            err
-                        ))
-                            .into()),
-                    }
-                } else {
-                    Ok(json["result"].clone())
-                }
-            }
-            Err(err) => Err(err.into())
-        }
-    }
-
-    pub async fn _send_request(&self, request_json: String) -> reqwest::Result<reqwest::Response> {
-        let mut too_many_requests_retries = 5;
-        loop {
-            let response = {
-                let client = self.client.clone();
-                let request_json = request_json.clone();
-                client
-                    .post(&self.url)
-                    .header(CONTENT_TYPE, "application/json")
-                    .body(request_json)
-                    .send()
-                    .await
-            };
-
-            match response {
-                Ok(response) => {
-                    if !response.status().is_success() {
-                        if response.status() == StatusCode::TOO_MANY_REQUESTS
-                            && too_many_requests_retries > 0
-                        {
-                            let mut duration = Duration::from_millis(500);
-                            if let Some(retry_after) = response.headers().get(RETRY_AFTER) {
-                                if let Ok(retry_after) = retry_after.to_str() {
-                                    if let Ok(retry_after) = retry_after.parse::<u64>() {
-                                        if retry_after < 120 {
-                                            duration = Duration::from_secs(retry_after);
-                                        }
-                                    }
-                                }
-                            }
-
-                            too_many_requests_retries -= 1;
-                            debug!(
-                                "Too many requests: server responded with {:?}, {} retries left, pausing for {:?}",
-                                response, too_many_requests_retries, duration
-                            );
-
-                            sleep(duration).await;
-                            continue;
-                        }
-                        return response.error_for_status();
-                    }
-                    return Ok(response);
-                }
-                other => return other
-            }
-        }
-    }
-
-    async fn get_evm_transaction_count(
-        &self,
-        address: &Address,
-    ) -> ClientResult<U256> {
-        self.send::<Hex<_>>(
-            RpcRequest::EthGetTransactionCount,
-            json!([evm_rpc::Hex(*address)]),
-        )
-            .await
-            .map(|h| h.0)
-    }
-
-    async fn get_evm_transaction_receipt(&self, hash: &H256) -> ClientResult<Option<RPCReceipt>> {
-        self.send::<Option<RPCReceipt>>(
-            RpcRequest::EthGetTransactionReceipt,
-            json!([evm_rpc::Hex(*hash)])
-        )
-            .await
-    }
-
-    async fn get_fee_calculator_for_blockhash_with_commitment(
-        &self,
-        blockhash: &Hash,
-        commitment_config: CommitmentConfig,
-    ) -> RpcResult<Option<FeeCalculator>> {
-        let RpcResponse { context, value } = self.send::<RpcResponse<Option<RpcFeeCalculator>>>(
-            RpcRequest::GetFeeCalculatorForBlockhash,
-            json!([
-                blockhash.to_string(),
-                commitment_config
-            ]),
-        )
-            .await?;
-
-        Ok(RpcResponse {
-            context,
-            value: value.map(|rf| rf.fee_calculator),
-        })
-    }
-
-    async fn get_signature_status_with_commitment(
-        &self,
-        signature: &Signature,
-        commitment_config: CommitmentConfig,
-    ) -> ClientResult<Option<solana_sdk::transaction::Result<()>>> {
-        let result: RpcResponse<Vec<Option<TransactionStatus>>> = self.send(
-            RpcRequest::GetSignatureStatuses,
-            json!([[signature.to_string()]]),
-        )
-            .await?;
-        Ok(result.value[0]
-            .clone()
-            .filter(|result| result.satisfies_commitment(commitment_config))
-            .map(|status_meta| status_meta.status))
-    }
-
-    async fn get_signature_status(
-        &self,
-        signature: &Signature,
-    ) -> ClientResult<Option<solana_sdk::transaction::Result<()>>> {
-        self.get_signature_status_with_commitment(
-            signature,
-            CommitmentConfig::processed()
-        ).await
-    }
-
-    async fn send_and_confirm_transaction_with_config(
-        &self,
-        transaction: &solana::Transaction,
-        config: RpcSendTransactionConfig,
-    ) -> ClientResult<Signature> {
-        const SEND_RETRIES: usize = 20;
-        const GET_STATUS_RETRIES: usize = 40;
-
-        'sending: for _ in 0..SEND_RETRIES {
-            let signature = self.send_transaction_with_config(transaction, config).await?;
-
-            let recent_blockhash = if uses_durable_nonce(transaction).is_some() {
-                let (recent_blockhash, ..) = self
-                    .get_recent_blockhash_with_commitment(CommitmentConfig::processed())
-                    .await?
-                    .value;
-                recent_blockhash
-            } else {
-                transaction.message.recent_blockhash
-            };
-
-            for status_retry in 0..GET_STATUS_RETRIES {
-                match self.get_signature_status(&signature).await? {
-                    Some(Ok(_)) => return Ok(signature),
-                    Some(Err(e)) => return Err(e.into()),
-                    None => {
-                        let fee_calculator = self
-                            .get_fee_calculator_for_blockhash_with_commitment(
-                                &recent_blockhash,
-                                CommitmentConfig::processed(),
-                            )
-                            .await?
-                            .value;
-                        if fee_calculator.is_none() {
-                            // Block hash is not found by some reason
-                            break 'sending;
-                        } else if cfg!(not(test))
-                            // Ignore sleep at last step.
-                            && status_retry < GET_STATUS_RETRIES
-                        {
-                            // Retry twice a second
-                            sleep(Duration::from_millis(500)).await;
-                            continue;
-                        }
-                    }
-                }
-            }
-        }
-
-        Err(RpcError::ForUser(
-            "unable to confirm transaction. \
-             This can happen in situations such as transaction expiration \
-             and insufficient fee-payer funds"
-                .to_string(),
-        )
-            .into())
-    }
-
-    async fn send_transaction_with_config(
-        &self,
-        transaction: &solana::Transaction,
-        config: RpcSendTransactionConfig,
-    ) -> ClientResult<Signature> {
-        let encoding = config.encoding.unwrap_or(UiTransactionEncoding::Base64);
-        let preflight_commitment = CommitmentConfig {
-            commitment: config.preflight_commitment.unwrap_or_default(),
-        };
-        let config = RpcSendTransactionConfig {
-            encoding: Some(encoding),
-            preflight_commitment: Some(preflight_commitment.commitment),
-            ..config
-        };
-        let serialized_encoded = serialize_encode_transaction(transaction, encoding)?;
-        let request = RpcRequest::SendTransaction;
-        let response = match self.send_request(request, json!([serialized_encoded, config])).await {
-            Ok(val) => serde_json::from_value(val)
-                .map_err(|err| ClientError::new_with_request(err.into(), request)),
-            Err(err) => Err(err.into_with_request(request)),
-        };
-        let signature_base58_str: String = match response {
-            Ok(signature_base58_str) => signature_base58_str,
-            Err(err) => {
-                if let ClientErrorKind::RpcError(RpcError::RpcResponseError {
-                                                     code,
-                                                     message,
-                                                     data,
-                                                     original_err: _original_err,
-                                                 }) = &err.kind
-                {
-                    debug!("{} {}", code, message);
-                    if let RpcResponseErrorData::SendTransactionPreflightFailure(
-                        RpcSimulateTransactionResult {
-                            logs: Some(logs), ..
-                        },
-                    ) = data
-                    {
-                        for (i, log) in logs.iter().enumerate() {
-                            debug!("{:>3}: {}", i + 1, log);
-                        }
-                        debug!("");
-                    }
-                }
-                return Err(err);
-            }
-        };
-
-        let signature = signature_base58_str
-            .parse::<Signature>()
-            .map_err(|err| Into::<ClientError>::into(RpcError::ParseError(err.to_string())))?;
-        // A mismatching RPC response signature indicates an issue with the RPC node, and
-        // should not be passed along to confirmation methods. The transaction may or may
-        // not have been submitted to the cluster, so callers should verify the success of
-        // the correct transaction signature independently.
-        if signature != transaction.signatures[0] {
-            Err(RpcError::RpcRequestError(format!(
-                "RPC node returned mismatched signature {:?}, expected {:?}",
-                signature, transaction.signatures[0]
-            ))
-                .into())
-        } else {
-            Ok(transaction.signatures[0])
-        }
-    }
-
-    async fn get_signature_statuses(
-        &self,
-        signatures: &[Signature],
-    ) -> RpcResult<Vec<Option<TransactionStatus>>> {
-        let signatures: Vec<_> = signatures.iter().map(|s| s.to_string()).collect();
-        let request = RpcRequest::GetSignatureStatuses;
-        let response = self
-            .send_request(request, json!([signatures]))
-            .await
-            .map_err(|err| err.into_with_request(request))?;
-        serde_json::from_value(response)
-            .map_err(|err| ClientError::new_with_request(err.into(), request))
-    }
-
-    async fn get_recent_blockhash_with_commitment(
-        &self,
-        commitment_config: CommitmentConfig,
-    ) -> RpcResult<(Hash, FeeCalculator, Slot)> {
-        let (context, blockhash, fee_calculator, last_valid_slot) = if let Ok(RpcResponse {
-            context,
-            value:
-                RpcFees {
-                    blockhash,
-                    fee_calculator,
-                    last_valid_slot,
-                    ..
-                },
-        }) = self.send::<RpcResponse<RpcFees>>(
-                RpcRequest::GetFees,
-                json!([commitment_config]),
-        )
-            .await
-        {
-            (context, blockhash, fee_calculator, last_valid_slot)
-        } else if let Ok(RpcResponse {
-            context,
-            value:
-                DeprecatedRpcFees {
-                    blockhash,
-                    fee_calculator,
-                    last_valid_slot,
-                },
-        }) = self.send::<RpcResponse<DeprecatedRpcFees>>(
-            RpcRequest::GetFees,
-            json!([commitment_config]),
-        )
-            .await
-        {
-            (context, blockhash, fee_calculator, last_valid_slot)
-        } else if let Ok(RpcResponse {
-            context,
-            value:
-                RpcBlockhashFeeCalculator {
-                    blockhash,
-                    fee_calculator,
-                },
-        }) = self.send::<RpcResponse<RpcBlockhashFeeCalculator>>(
-            RpcRequest::GetRecentBlockhash,
-            json!([commitment_config]),
-        )
-            .await
-        {
-            (context, blockhash, fee_calculator, 0)
-        } else {
-            return Err(ClientError::new_with_request(
-                RpcError::ParseError("RpcBlockhashFeeCalculator or RpcFees".to_string()).into(),
-                RpcRequest::GetRecentBlockhash,
-            ));
-        };
-
-        let blockhash = blockhash.parse().map_err(|_| {
-            ClientError::new_with_request(
-                RpcError::ParseError("Hash".to_string()).into(),
-                RpcRequest::GetRecentBlockhash,
-            )
-        })?;
-        Ok(RpcResponse {
-            context,
-            value: (blockhash, fee_calculator, last_valid_slot),
-        })
-    }
-
-    async fn get_recent_blockhash(&self) -> ClientResult<(Hash, FeeCalculator)> {
-        let (blockhash, fee_calculator, _last_valid_slot) = self
-            .get_recent_blockhash_with_commitment(CommitmentConfig::processed())
-            .await?
-            .value;
-        Ok((blockhash, fee_calculator))
-    }
-
-    async fn get_new_blockhash(&self, blockhash: &Hash) -> ClientResult<(Hash, FeeCalculator)> {
-        let mut num_retries = 0;
-        let start = Instant::now();
-        while start.elapsed().as_secs() < 5 {
-            if let Ok((new_blockhash, fee_calculator)) = self.get_recent_blockhash().await {
-                if new_blockhash != *blockhash {
-                    return Ok((new_blockhash, fee_calculator));
-                }
-            }
-            debug!("Got same blockhash ({:?}), will retry...", blockhash);
-
-            // Retry ~twice during a slot
-            sleep(Duration::from_millis(DEFAULT_MS_PER_SLOT / 2)).await;
-            num_retries += 1;
-        }
-        Err(RpcError::ForUser(format!(
-            "Unable to get new blockhash after {}ms (retried {} times), stuck at {}",
-            start.elapsed().as_millis(),
-            num_retries,
-            blockhash
-        ))
-            .into())
-    }
-
-    async fn get_minimum_balance_for_rent_exemption(&self, data_len: usize) -> ClientResult<u64> {
-        let request = RpcRequest::GetMinimumBalanceForRentExemption;
-        let minimum_balance_json = self
-            .send_request(request, json!([data_len]))
-            .await
-            .map_err(|err| err.into_with_request(request))?;
-
-        let minimum_balance: u64 = serde_json::from_value(minimum_balance_json)
-            .map_err(|err| ClientError::new_with_request(err.into(), request))?;
-        trace!(
-            "Response minimum balance {:?} {:?}",
-            data_len,
-            minimum_balance
-        );
-        Ok(minimum_balance)
-    }
-
-    async fn send<T>(&self, request: RpcRequest, params: Value) -> ClientResult<T>
-        where
-            T: serde::de::DeserializeOwned,
-    {
-        assert!(params.is_array() || params.is_null());
-        let response = self
-            .send_request(request, params)
-            .await
-            .map_err(|err| err.into_with_request(request))?;
-        serde_json::from_value(response)
-            .map_err(|err| ClientError::new_with_request(err.into(), request))
-    }
 }
 
 #[derive(Derivative)]
@@ -806,7 +325,7 @@ impl EvmBridge {
 
         match is_receipt_exists(self, hash).await {
             Some(b) => Some(b),
-            None => is_signature_exists(self, hash).await
+            None => is_signature_exists(self, hash).await,
         }
     }
 }
@@ -861,7 +380,8 @@ impl BridgeERPC for BridgeErpcImpl {
             let nonce = match tx
                 .nonce
                 .map(|a| a.0)
-                .or_else(|| meta.pool.transaction_count(&address)) {
+                .or_else(|| meta.pool.transaction_count(&address))
+            {
                 Some(n) => n,
                 None => meta
                     .rpc_client
@@ -918,7 +438,8 @@ impl BridgeERPC for BridgeErpcImpl {
             let nonce = match tx
                 .nonce
                 .map(|a| a.0)
-                .or_else(|| meta.pool.transaction_count(&address)) {
+                .or_else(|| meta.pool.transaction_count(&address))
+            {
                 Some(n) => n,
                 None => meta
                     .rpc_client
@@ -1080,14 +601,7 @@ impl ChainERPC for ChainErpcProxy {
         address: Hex<Address>,
         block: Option<BlockId>,
     ) -> BoxFuture<EvmResult<Hex<U256>>> {
-        Box::pin(async move {
-            proxy_evm_rpc!(
-                meta.rpc_client,
-                EthGetBalance,
-                address,
-                block
-            )
-        })
+        Box::pin(async move { proxy_evm_rpc!(meta.rpc_client, EthGetBalance, address, block) })
     }
 
     #[instrument(skip(self))]
@@ -1098,15 +612,9 @@ impl ChainERPC for ChainErpcProxy {
         data: Hex<U256>,
         block: Option<BlockId>,
     ) -> BoxFuture<EvmResult<Hex<H256>>> {
-        Box::pin(async move {
-            proxy_evm_rpc!(
-                meta.rpc_client,
-                EthGetStorageAt,
-                address,
-                data,
-                block
-            )
-        })
+        Box::pin(
+            async move { proxy_evm_rpc!(meta.rpc_client, EthGetStorageAt, address, data, block) },
+        )
     }
 
     #[instrument(skip(self))]
@@ -1122,14 +630,9 @@ impl ChainERPC for ChainErpcProxy {
             }
         }
 
-        Box::pin(async move {
-            proxy_evm_rpc!(
-                meta.rpc_client,
-                EthGetTransactionCount,
-                address,
-                block
-            )
-        })
+        Box::pin(
+            async move { proxy_evm_rpc!(meta.rpc_client, EthGetTransactionCount, address, block) },
+        )
     }
 
     #[instrument(skip(self))]
@@ -1139,11 +642,7 @@ impl ChainERPC for ChainErpcProxy {
         block: BlockId,
     ) -> BoxFuture<EvmResult<Hex<usize>>> {
         Box::pin(async move {
-            proxy_evm_rpc!(
-                meta.rpc_client,
-                EthGetBlockTransactionCountByNumber,
-                block
-            )
+            proxy_evm_rpc!(meta.rpc_client, EthGetBlockTransactionCountByNumber, block)
         })
     }
 
@@ -1169,14 +668,7 @@ impl ChainERPC for ChainErpcProxy {
         address: Hex<Address>,
         block: Option<BlockId>,
     ) -> BoxFuture<EvmResult<Bytes>> {
-        Box::pin(async move {
-            proxy_evm_rpc!(
-                meta.rpc_client,
-                EthGetCode,
-                address,
-                block
-            )
-        })
+        Box::pin(async move { proxy_evm_rpc!(meta.rpc_client, EthGetCode, address, block) })
     }
 
     #[instrument(skip(self))]
@@ -1190,12 +682,7 @@ impl ChainERPC for ChainErpcProxy {
             Box::pin(ready(Ok(Some(RPCBlock::default()))))
         } else {
             Box::pin(async move {
-                proxy_evm_rpc!(
-                    meta.rpc_client,
-                    EthGetBlockByHash,
-                    block_hash,
-                    full
-                )
+                proxy_evm_rpc!(meta.rpc_client, EthGetBlockByHash, block_hash, full)
                     .map(|o: Option<_>| o.map(compatibility::patch_block))
             })
         }
@@ -1212,12 +699,7 @@ impl ChainERPC for ChainErpcProxy {
             Box::pin(ready(Ok(Some(RPCBlock::default()))))
         } else {
             Box::pin(async move {
-                proxy_evm_rpc!(
-                    meta.rpc_client,
-                    EthGetBlockByNumber,
-                    block,
-                    full
-                )
+                proxy_evm_rpc!(meta.rpc_client, EthGetBlockByNumber, block, full)
                     .map(|o: Option<_>| o.map(compatibility::patch_block))
             })
         }
@@ -1237,11 +719,7 @@ impl ChainERPC for ChainErpcProxy {
             }
         }
         Box::pin(async move {
-            proxy_evm_rpc!(
-                meta.rpc_client,
-                EthGetTransactionByHash,
-                tx_hash
-            )
+            proxy_evm_rpc!(meta.rpc_client, EthGetTransactionByHash, tx_hash)
                 .map(|o: Option<_>| o.map(compatibility::patch_tx))
         })
     }
@@ -1286,13 +764,7 @@ impl ChainERPC for ChainErpcProxy {
         meta: Self::Metadata,
         tx_hash: Hex<H256>,
     ) -> BoxFuture<EvmResult<Option<RPCReceipt>>> {
-        Box::pin(async move {
-            proxy_evm_rpc!(
-                meta.rpc_client,
-                EthGetTransactionReceipt,
-                tx_hash
-            )
-        })
+        Box::pin(async move { proxy_evm_rpc!(meta.rpc_client, EthGetTransactionReceipt, tx_hash) })
     }
 
     #[instrument(skip(self))]
@@ -1303,15 +775,7 @@ impl ChainERPC for ChainErpcProxy {
         block: Option<BlockId>,
         meta_keys: Option<Vec<String>>,
     ) -> BoxFuture<EvmResult<Bytes>> {
-        Box::pin(async move {
-            proxy_evm_rpc!(
-                meta.rpc_client,
-                EthCall,
-                tx,
-                block,
-                meta_keys
-            )
-        })
+        Box::pin(async move { proxy_evm_rpc!(meta.rpc_client, EthCall, tx, block, meta_keys) })
     }
 
     #[instrument(skip(self))]
@@ -1322,15 +786,9 @@ impl ChainERPC for ChainErpcProxy {
         block: Option<BlockId>,
         meta_keys: Option<Vec<String>>,
     ) -> BoxFuture<EvmResult<Hex<Gas>>> {
-        Box::pin(async move {
-            proxy_evm_rpc!(
-                meta.rpc_client,
-                EthEstimateGas,
-                tx,
-                block,
-                meta_keys
-            )
-        })
+        Box::pin(
+            async move { proxy_evm_rpc!(meta.rpc_client, EthEstimateGas, tx, block, meta_keys) },
+        )
     }
 
     #[instrument(skip(self, meta))]
@@ -1691,13 +1149,15 @@ async fn send_and_confirm_transactions<T: Signers>(
             let mut retained = vec![];
             for (transaction, signature) in transactions_signatures {
                 if let Some(signature) = signature {
-                    if rpc_client.get_signature_statuses(&[signature])
+                    if rpc_client
+                        .get_signature_statuses(&[signature])
                         .await
                         .ok()
                         .and_then(|RpcResponse { mut value, .. }| value.remove(0))
                         .and_then(|status| status.confirmations)
                         .map(|confirmations| confirmations == 0) // retain unconfirmed only
-                        .unwrap_or(true) {
+                        .unwrap_or(true)
+                    {
                         retained.push((transaction, Some(signature)));
                     }
                 } else {
@@ -1713,7 +1173,8 @@ async fn send_and_confirm_transactions<T: Signers>(
 
         // Re-sign any failed transactions with a new blockhash and retry
         let (blockhash, _) = rpc_client
-            .get_new_blockhash(&transactions_signatures[0].0.message().recent_blockhash).await?;
+            .get_new_blockhash(&transactions_signatures[0].0.message().recent_blockhash)
+            .await?;
 
         for (mut transaction, _) in transactions_signatures {
             transaction.try_sign(signer_keys, blockhash)?;
@@ -1726,7 +1187,7 @@ async fn send_and_confirm_transactions<T: Signers>(
 
 #[cfg(test)]
 mod tests {
-    use crate::{BridgeErpcImpl, EthPool, EvmBridge, AsyncRpcClient, SystemClock};
+    use crate::{AsyncRpcClient, BridgeErpcImpl, EthPool, EvmBridge, SystemClock};
     use evm_rpc::{BridgeERPC, Hex};
     use evm_state::Address;
     use secp256k1::SecretKey;
